Consolidate isDateFilterIrrelevant and export it
There were to versions of this function.
Also, we need it outside, so it is now in the barrel.

JIRA: RAIL-2847
