+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/KpiAlerts/utils/brokenFilterUtils.ts
@@ -26,7 +26,7 @@ import last from "lodash/last";
 import partition from "lodash/partition";
 import { IntlShape } from "react-intl";
 
-import { isAttributeFilterIgnored, isDateFilterIrrelevant } from "../utils/filterUtils";
+import { isAttributeFilterIgnored, isDateFilterIrrelevant } from "../../utils/filters";
 
 import { dashboardDateFilterToPreset } from "./translationUtils";
 import {
